Add files via upload
In this project, we will make some list items with HTML elements that can be dragged and dropped to another block using HTML and CSS. We can drag the HTML element from block 1 to block 2 and block 2 to block 1 with JavaScript. Here we will use JavaScript events called "dragstart", "dragover" and "Drop" events.